+++ book-visit-app-server/.env
@@ -1,5 +1,7 @@
 NODE_SERVER_IP=http://localhost:5000
 NODE_SERVER_PORT=5000
+CLIENT_IP=http://localhost:3000
+CLIENT_PORT=3000
 
 NODE_GMAIL_SENDER_MAIL=dawid.book.app@gmail.com
 NODE_GMAIL_SENDER_PASSWORD=zaq1@WSXQwEr!
