+++ book-visit-app-server/views/accountActivated.html
@@ -1,0 +1,65 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="icon" href="../public/images/logo/favicon.png" />
+    <title>Success in activating account</title>
+    <style>
+        * {
+            margin: 0;
+            padding: 0;
+            box-sizing: border-box;
+            font-family: sans-serif;
+        }
+
+        .container {
+            position: relative;
+            width: 960px;
+            height: 100vh;
+            left: calc(50% - 480px);
+            padding: 70px;
+            background-color: #FFFDD0;
+        }
+
+        h2 {
+            text-align: center;
+            color: royalblue;
+            font-weight: bold;
+            margin: 30px 0 50px;
+            font-size: 38px;
+        }
+
+        button {
+            position: relative;
+            left: calc(50% - 100px);
+            height: 60px;
+            width: 200px;
+            background-color: royalblue;
+            color: white;
+            font-weight: bold;
+            /* border: 2px solid darkblue; */
+            box-shadow: darkblue 1px 1px 20px;
+            border-radius: 5px;
+        }
+
+        a {
+            text-decoration: none;
+            color: white;
+            font-weight: bold;
+            font-size: 16px;
+        }
+    </style>
+</head>
+
+<body>
+    <div class="container">
+        <h2>Your accont has been successfully activated!</h2>
+        <button><a href="http://127.0.0.1:3000/login">Go back to login page</a></button>
+    </div>
+
+</body>
+
+</html>
