+++ book-visit-app/src/components/App.js
@@ -1,9 +1,10 @@
 import React, { useState, useEffect } from "react";
 import "./App.css";
-import Orders from "./Orders";
-import LoginContainer from "./LoginContainter";
-import AddOrder from "./AddOrder";
-import MyOrders from "./MyOrders";
+import Orders from "./orders/Orders";
+import LoginContainer from "./auth/LoginContainter";
+import AddOrder from "./orders/AddOrder";
+import MyOrders from "./orders/MyOrders";
+import OrderDetail from './orders/OrderDetail'
 import { UserContext } from "../contexts/UserContext";
 
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
@@ -18,6 +19,7 @@ function App() {
           <Route exact path="/" component={Home}></Route>
           <Route path="/login" component={LoginContainer}></Route>
           <Route path="/orders" component={Orders}></Route>
+          <Route path="/order/:orderId" component={OrderDetail}></Route>
           <Route path="/myorders/:workerId" component={MyOrders}></Route>
           <Route path="/addorder" component={AddOrder}></Route>
         </Switch>
