+++ book-visit-app/src/components/orders/OrderDetail.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const OrderDetail = (props) => {
+    return (
+        <div>
+
+        </div>
+    )
+}
+
+export default OrderDetail
